+++ dart_ping/lib/dart_ping.dart
@@ -1,7 +1,8 @@
 library;
 
 export 'package:dart_ping/src/models/ping_event.dart';
-export 'package:dart_ping/src/models/round_trip_stats.dart';
+export 'package:dart_ping/src/models/round_trip_stats.dart'
+    show RoundTripStats, RoundTripStatsAccumulator;
 export 'package:dart_ping/src/ip_version.dart';
 export 'package:dart_ping/src/address_family.dart'
     show ipLiteralFamily, validateAddressFamily;